refactor(migration): move addTableColumnSearchSettings from v200 to v1130
The tableColumn entity type was introduced in 1.13.0; the migration that
registers it as a searchable asset belongs with the version that introduced
it, not with 2.0.0 which carries the unrelated Task Redesign migration.

This is purely an attribution / packaging change:
  - Move addTableColumnSearchSettings() and TABLE_COLUMN_ASSET_TYPE constant
    from v200/MigrationUtil.java to v1130/MigrationUtil.java, with related
    SearchSettings/Settings/SearchSettingsMergeUtil imports shifted along.
  - Wire MigrationUtil.addTableColumnSearchSettings() into both Postgres and
    MySQL v1130/Migration.java#runDataMigration() (idempotent helper, safe
    on every reprocessing pass per PR #26571 contract).
  - Remove the call and static import from both v200/Migration.java files;
    drop the now-unused SearchSettings imports from v200/MigrationUtil.java.

Behaviour:
  - Fresh installs: 1.13.0 migration registers tableColumn (where it belongs).
  - Existing 2.0.0+ deploys: the helper is idempotent — detects the
    assetTypeConfiguration / allowedFields entries already exist and no-ops.
  - 2.0.0 migration retains its Task Redesign work, which is the legitimate
    2.0.0 scope.
