+++ openmetadata-service/src/main/java/org/openmetadata/service/migration/postgres/v200/Migration.java
@@ -1,7 +1,6 @@
 package org.openmetadata.service.migration.postgres.v200;
 
 import static org.openmetadata.service.jdbi3.locator.ConnectionType.POSTGRES;
-import static org.openmetadata.service.migration.utils.v200.MigrationUtil.addTableColumnSearchSettings;
 import static org.openmetadata.service.migration.utils.v200.MigrationUtil.backfillAnnouncementRelationships;
 import static org.openmetadata.service.migration.utils.v200.MigrationUtil.migrateLegacyActivityThreadsToActivityStream;
 import static org.openmetadata.service.migration.utils.v200.MigrationUtil.migrateSuggestionsToTaskEntity;
@@ -20,7 +19,6 @@ public class Migration extends MigrationProcessImpl {
   @Override
   @SneakyThrows
   public void runDataMigration() {
-    addTableColumnSearchSettings();
     migrateSuggestionsToTaskEntity(handle, POSTGRES);
     migrateThreadTasksToTaskEntity(handle, POSTGRES);
     migrateLegacyActivityThreadsToActivityStream(handle, POSTGRES);
